sep: free_all
