+++ libft/Makefile
@@ -66,7 +66,8 @@ PUT_SRC		:=	$(PUT_DIR)/ft_putchar_fd.c \
 # ft_string
 #--------------------------------------------
 STRING_DIR	:=	ft_string
-STRING_SRC	:=	$(STRING_DIR)/ft_split.c \
+STRING_SRC	:=	$(STRING_DIR)/ft_free.c \
+				$(STRING_DIR)/ft_split.c \
 				$(STRING_DIR)/ft_strchr.c \
 				$(STRING_DIR)/ft_strdup.c \
 				$(STRING_DIR)/ft_striteri.c \
