+++ srcs/main.c
@@ -3,20 +3,6 @@
 #include "minishell.h"
 #include "ft_string.h"
 
-void	*free_all(char **split_strs)
-{
-	size_t	i;
-
-	i = 0;
-	while (split_strs[i] != NULL)
-	{
-		free(split_strs[i]);
-		i++;
-	}
-	free(split_strs);
-	return (NULL);
-}
-
 int	minishell(void)
 {
 	char		*line;
